requirements.txt
pandas==1.3.3
scikit-learn==0.24.2
joblib==1.0.1
hydra-core==1.1.1
omegaconf==2.1.1
pytest==6.2.5
fastapi==0.68.1
uvicorn==0.15.0
motor==2.5.1
pymongo==3.12.0
python-dotenv==0.19.0
requests==2.26.0
matplotlib==3.4.3
seaborn==0.11.2
xgboost==1.4.2
